created useGetHearingSlotMetaData
