+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/package.json
@@ -18,7 +18,6 @@
     "react-router-dom": "5.3.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.24.7",
     "@cyntler/react-doc-viewer": "1.12.0",
     "@egovernments/digit-ui-components": "0.0.1-beta.28",
     "@egovernments/digit-ui-libraries": "1.8.1-beta.2",
